add middleware that doesnt work
